+++ config/logging.php
@@ -50,7 +50,7 @@ return [
     'channels' => [
         'stack' => [
             'driver' => 'stack',
-            'channels' => ['single'],
+            'channels' => ['single', 'discord'],
             'ignore_exceptions' => false,
         ],
 
@@ -112,6 +112,20 @@ return [
 
         'emergency' => [
             'path' => storage_path('logs/laravel.log'),
+        ],
+
+        'discord' => [
+          'driver' => 'custom',
+          'via'    => MarvinLabs\DiscordLogger\Logger::class,
+          'level'  => 'debug',
+          'url'    => env('LOG_DISCORD_WEBHOOK_URL'),
+        ],
+
+        'discord_errors' => [
+          'driver' => 'custom',
+          'via'    => MarvinLabs\DiscordLogger\Logger::class,
+          'level'  => 'debug',
+          'url'    => env('LOG_DISCORD_ERRORS_WEBHOOK_URL'),
         ],
     ],