+++ resources/lang/en/mail.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+
+  /*
+  |--------------------------------------------------------------------------
+  | Mailer Language Lines
+  |--------------------------------------------------------------------------
+  |
+  | The following language lines are used during mailing for various
+  | messages that we need to display to the user. You are free to modify
+  | these language lines according to your application's requirements.
+  |
+  */
+
+  'Welcome' => 'Welcome',
+  'Welcome Paragraph' => 'Thanks for signing up',
+  'Welcome Title' => 'Welcome to Finch'
+
+];