+++ routes/console_routes/conduit_routes.php
@@ -1,7 +1,10 @@
 <?php
 
+use App\Mail\TestMail;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\Log;
 
 /*
 |--------------------------------------------------------------------------
@@ -29,4 +32,41 @@ Artisan::command('conduit:check-connection', function () {
     die;
   }
 })->purpose('Checks the status of the Conduit Database Connection');
+
+
+
+/**
+ * Test Email System
+ */
+Artisan::command('conduit:check-email', function () {
+  $this->comment('Checking the Conduit Email System');
+
+  $test_recipient = 'me@benbroad.com';
+
+  Mail::to($test_recipient)->send(new TestMail('Benjamin Number 2'));
+  Log::info('A test email has been sent to ' . $test_recipient);
+
+})->purpose('Sends an email to test the Conduit Email System');
+
+
+/**
+ * Test Discord Logger
+ */
+Artisan::command('conduit:check-discord', function () {
+  $this->comment('Checking the Conduit Discord Logger');
+
+  Log::channel('discord')->info('Discord Logger is working as expected.');
+
+})->purpose('Sends an notification to test the Conduit Discord Logger');
+
+
+/**
+ * Test Discord Error Logger
+ */
+Artisan::command('conduit:check-discord-error', function () {
+  $this->comment('Checking the Conduit Discord Error Logger');
+
+  Log::channel('discord_errors')->warning('Discord Error Logger is working as expected.');
+
+})->purpose('Sends an notification to test the Conduit Discord Error Logger');